+++ classifiers/behavior.py
@@ -1,0 +1,69 @@
+"""file containing functions for classifying behavior."""
+
+import numpy as np
+import core.config as config
+from collections import deque
+
+
+def update_joint_history(person_id, skeleton):
+    if person_id not in config.joint_histories:
+        config.joint_histories[person_id] = {
+            joint: deque(maxlen=config.WINDOW_SIZE) for joint in config.FIDGET_JOINTS
+        }
+    for joint in config.FIDGET_JOINTS:
+        kp = skeleton.keypoint[joint]
+        if not np.allclose(kp, [0, 0, 0]):
+            config.joint_histories[person_id][joint].append(np.array(kp))
+
+
+def compute_velocity(positions):
+    velocities = []
+    for i in range(1, len(positions)):
+        v = np.linalg.norm(positions[i] - positions[i - 1])
+        velocities.append(v)
+    return np.array(velocities)
+
+
+def classify_fidgeting(joint_history):
+    motion_energies = []
+    for joint, positions in joint_history.items():
+        if len(positions) < 2:
+            continue
+        positions_np = np.array(positions)
+        velocities = compute_velocity(positions_np)
+        above_thresh = velocities > config.MIN_MOTION_THRESHOLD
+        if np.sum(above_thresh) >= config.MIN_ACTIVE_FRAMES:
+            motion_energies.append(np.mean(velocities))  # or np.sum(velocities)
+
+    if not motion_energies:
+        return 0.0
+    return np.mean(motion_energies)
+
+
+def get_behavior(person):
+    pid = person.id
+    action = person.action_state
+    update_joint_history(pid, person)
+
+    if str(action) == "MOVING":
+        return "moving"
+
+    elif str(action) == "IDLE":
+        if pid not in config.joint_histories:
+            return "still"  # Not enough data yet, assume still
+
+        score = classify_fidgeting(config.joint_histories[pid])
+
+        if score != 0.0:
+            norm_score = np.clip((score - 0.0) / (0.1 - 0.0), 0, 1)
+            if norm_score < 0.2:
+                return "slightly fidgety"
+            elif norm_score < 0.5:
+                return "fidgety"
+            else:
+                return "very fidgety"
+        else:
+            return "still"
+
+    else:
+        return "unknown"
